pyproject.toml
[build-system]
requires      = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "garak"
version = "0.11.0.pre1"
authors = [
  { name = "Leon Derczynski", email="lderczynski@nvidia.com" },
  { name = "Subho Majumdar", email="subho@vijil.ai" },
  { name = "nv052193" },
  { name = "Mads Kongsbak" },
  { name = "Erick Galinkin", email="egalinkin@nvidia.com" },
  { name = "Tianhao Li" },
  { name = "Phyllis Poh" },
  { name = "Razvan Dinu" },
  { name = "Zander Mackie" },
  { name = "Greg Stephens" },
  { name = "Jeffrey Martin", email="jemartin@nvidia.com" },
  { name = "Ahsan Ayub" },
  { name = "Jonathan Liberman" },
  { name = "Gustav Fredrikson" },
  { name = "Oh Tien Cheng" },
  { name = "Brain John" },
  { name = "Naman Mishra" },
  { name = "Soumili Nandi" },
  { name = "Arjun Krishna" },
  { name = "Mihailo Milenkovic" },
  { name = "Kai Greshake" },
  { name = "Martin Borup-Larsen" },
  { name = "Emmanuel Ferdman" },
  { name = "Eric Therond" },
  { name = "Zoe Nolan" },
  { name = "Harsh Raj" },
  { name = "Shine-afk" },
  { name = "Rafael Sandroni" },  
  { name = "Eric Hacker" },
  { name = "Blessed Uyo" },
  { name = "Ikko Eltociear Ashimine" },
  { name = "iamnotcj" },
  { name = "Dwight Temple" },
  { name = "Shane Rosse" },
  { name = "Masaya Ogushi" },
]
license = { file = "LICENSE" }
description = "LLM vulnerability scanner"
readme = "README.md"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: Apache Software License",
]
requires-python = ">=3.10"
dependencies = [
  "base2048>=0.1.3",
  "transformers>=4.43.0",
  "datasets>=3.0.0",
  "colorama>=0.4.3",
  "tqdm>=4.64.0",
  "cohere>=4.5.1,<5",
  "openai>=1.45.0,<2",
  "replicate>=0.8.3",
  "google-api-python-client>=2.0",
  "backoff>=2.1.1",
  "rapidfuzz>=3.0.0",
  "jinja2>=3.1.2",
  "nltk>=3.8.1",
  "accelerate>=0.23.0",
  "avidtools==0.1.2",
  "stdlibs>=2022.10.9",
  "langchain>=0.0.300",
  "nemollm>=0.3.0",
  "octoai-sdk>=0.8.0",
  "cmd2==2.4.3",
  "torch>=2.1.3",
  "sentencepiece>=0.1.99",
  "markdown>=3.4.3",
  "numpy>=1.26.1",
  "zalgolib>=0.2.2",
  "ecoji>=0.1.1",
  "deepl==1.17.0",
  "fschat>=0.2.36",
  "litellm>=1.41.21,<1.67.2",
  "jsonpath-ng>=1.6.1",
  "huggingface_hub>=0.21.0",
  'python-magic-bin>=0.4.14; sys_platform == "win32"',
  'python-magic>=0.4.21; sys_platform != "win32"',
  "lorem==0.1.1",
  "xdg-base-dirs>=6.0.1",
  "wn==0.9.5",
  "ollama>=0.4.7",
  "nvidia-riva-client==2.16.0",
  "google-cloud-translate>=2.0.4",
  "langdetect==1.0.9",
  "tiktoken>=0.7.0",
  "mistralai==1.5.2",
  "ftfy>=6.3.1",
]

[project.optional-dependencies]
tests = [
  "pytest>=8.0",
  "pytest-mock>=3.14.0",
  "requests-mock==1.12.1",
  "respx>=0.21.1",
  "pytest-cov>=5.0.0",
  "pytest_httpserver>=1.1.0",
  "langcodes>=3.4.0",
]
lint = [
  "black==24.4.2",
  "pylint>=3.1.0",
]
calibration = [
  "scipy>=1.14.0",
]

[project.urls]
"Homepage" = "https://github.com/NVIDIA/garak"
"Bug Tracker" = "https://github.com/NVIDIA/garak/issues"

[project.scripts]
garak = "garak.__main__:main"

[tool.setuptools]
packages = ["garak"]

[tool.setuptools.package-data]
"garak.resources" = ["*"]

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'

[tool.pytest.ini_options]
filterwarnings = [
    "ignore",
    "default:::garak",
]

[tool.coverage.run]
source = ["./garak"]
omit = ["tests/*"]
